First implementation of resource lookup
Resources provide easy ways to access the Pinterest API functionality:
they work as permissive aliases to the endpoints.
These are intented to be much easier to figure out.
For example, compare the endpoint

    DELETE /v1/me/following/boards/:board/

vs

    delete('my/following/board', ...

or

    GET /v1/me/search/pins/

vs

    fetch('search/my/pins', ...

They are also insensitive to singular/plural for words like
"boards", "user", "pins', etc. and insensitive to trailing slashes.